Omnijaws: widget: Uncouple the temperature from the image
Until this change, the temperature strings were combined together with the weather condition images. This caused the temperature string to be clipped when using some icon packs.

How it looks: https://imgur.com/a/Kj3w3z3

Let's just uncouple the text and the image and set the text to a separate Textview. Also improves readability
Change-Id: I4f43fa8d36575ecd69ab7829c184625a76956f92

Change-Id: I48e9216aec437df8c09d497bc7118df4f8956939